Cleaned up Python code a little bit
